requirements.txt
ansible==2.10.4
ansible-base==2.10.4
ansible-runner==1.4.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
docutils==0.16
idna==2.10
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
packaging==20.8
pexpect==4.8.0
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pyparsing==2.4.7
python-daemon==2.2.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2